Snowflake sample code added & type of id change from int to long

+++ src/main/java/com/m/blog/domain/boardCollection/infrastructure/repository/BoardAggregationDto.java
@@ -7,15 +7,15 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class BoardAggregationDto {
-    int boardCollectionId;
+    long boardCollectionId;
     String boardCollectionName;
-    int boardId;
+    long boardId;
     String boardName;
     long postingCount;
 
     @QueryProjection
-    public BoardAggregationDto(int boardCollectionId, String boardCollectionName,
-                               int boardId, String boardName, long postingCount) {
+    public BoardAggregationDto(long boardCollectionId, String boardCollectionName,
+                               long boardId, String boardName, long postingCount) {
         this.boardCollectionId = boardCollectionId;
         this.boardCollectionName = boardCollectionName;
         this.boardId = boardId;
